Relocate `StandaloneTagLibraryLookup` from `grace-web-gsp-taglib` to `grace-web-taglib`
